requirements.txt
pygame>=2.1.0
neat-python>=0.92.0
numpy>=1.23.0